use getJustDateFromUTCDateString & move dates file
